Подключает данные с сервера к проекту
