+++ js/editing-picture.js
@@ -211,4 +211,3 @@ const REMOVE_LISTENER_EFFECTS = function () {
 };
 
 export {ZOOM_INITIAL, ADD_LISTENER_SCALES_CONTROL, REMOVE_LISTENER_SCALES_CONTROL, INIT_LEVEL_SLIDER, ADD_LISTENER_EFFECTS, REMOVE_LISTENER_EFFECTS};
-// module9-task2
