+++ js/main.js
@@ -1,5 +1,7 @@
 import './generate-photo-gallery/generate-photo-gallery.js';
 import './rendering-pictures.js';
 import './big-picture.js';
-import './form.js';
+import './user-form.js'; // Почему работает без импорта?
+import './validate-form.js';
 import './editing-picture.js';
+import './api.js';
